RegFieldDesc: move to a separate file
